Create decks/stats.rs
